Hello World Commit
Created a hello world program with OpenGL that creates a window and set a certain color. When the user closes the window, the program exits.

Had to research plenty about how to get started with CMake, link C++ libraries, and build a library using a source code of public repository.